Dockerfile: install "tini" as init and set as ENTRYPOINT
The README suggests

  sudo podman run --pull always --replace --pid host --network host --user 0 --name marvell-tools -dit --privileged -v /dev:/dev quay.io/sdaniele/marvell-tools
  sudo podman exec -it marvell-tools <cmd>

In this case, the pod is running /bin/bash, which does not terminate on
SIGTERM. That means, `podman stop` will hang.

Fix that, by instead using "sleep infinity". That process will terminate
on SIGTERM and no leave the pod hanging.

While at it, also use tini as init system. For our use case the overhead
of an additional process does not matter, and having a init that
delivers signals can be useful in some situations. Set the ENTRYPOINT to
use tini, while CMD to an infinite sleep.

For this we need to enable EPEL, which seems anyway useful to have on
CentOS Stream 9.

What the best ENTRYPOINT/CMD is, depends on the command that we want to
call and how we run the container (using podman? using --pid host?,
using --init?). There are various considerations that depend on the use
case, but overall I think this is the best way.