pxeboot: create NetworkManager connection profiles on DPU during kickstart
The interface "enP2p3s0" inside the DPU is expected to be connected to
"eno4" on the host. It gets a NetworkManger connection profile
"enP2p3s0-dpu-host" with a static IP address, alongside running
DHCP/SLAAC. The static configuration is useful, if the host no longer
has a DHCP server running on "eno4".

The "enP2p2s0" interface is expected to the provisioning host, which may
run DHCP there. Add a NetworkManager profile "enP2p2s0-dpu-secondary"
which does DHCP/SLAAC.