all: don't rely on current working directory to find files
As it was, the files were only found if the current working directory is
the one with the "manifests/" directory.

Which usually happens to be the case, if we start `podman` without "-w"
option and place the files to "/manifests". However, the location of
those files will change. Also, the code should work regardless of the
current working directory.