pxeboot: add "--host-setup-only" and "--host-mode" options
The pxeboot.py command also needs to configure a few things on the host.
At the very least, it needs to configure 172.131.100.1/24 address on
"eno4" interface (either via iproute2 or persistently via a
NetworkManager connection profile). It probably also should always
configure the nft rules for NAT.

If the host is running RHEL, we are expected that the user "works" on
this host. In that case, we want to perform persistent configuration
(via a NetworkManager connnection profile). If the host is running
CoreOS, changing the host makes little sense. Add a "--host-mode" flag
that controls how the host gets modified.

Also, the tool performs ephemeral configuration (nft rules and
net.ipv4.ip_forward sysctl). That is, because there is no convenient API
for making that configuration persistent, and on CoreOS it wouldn't make
sense anyway. That means, if you reboot the host, the nft rules and
sysfs settings are lost. Add a "--host-setup-only" option. When this
flag is set, the tool does nothing to the DPU, but performs the
configuration of the host only. This is useful after a reboot.