README: in usage examples use IMAGE variable
A main point of the usage examples in the README is to be able to
copy+paste the snippets. Move the selection of the image to a
IMAGE variable, so that we can easier copy+paste and select another
image.

Also, use the same image in all examples. It is the one that is
currently build by the github workflows.