pxeboot,fwupdate: remove kill_existing() to kill other pxeboot.py processes
The way kill_existing() is implemented seems very ugly. We identify a
"pxeboot.py" process by reading /proc and send it SIGKILL. What this in
particular does, is to keep the dhcpd service running. The code does
something, but it's unclear whether this is at all the best way.

The best way to run the tool is following the usage example in the README.md,
which does a `podman run --rm --replace --name marvell-tools...`.
If you do that, podman takes care that no concurrent dhcpd processes are
running and will kill other instances before starting a new one.

If you don't follow that, then the tool cannot magically know which
processes on the system might interfere and should be stopped. The
caller must ensure that the system does not have conflicting services
running, and using `podman run` as suggested helps with that.

This code seems unnecessary. Remove it.