pxeboot,fwupdate: remove kill_existing() to kill other pxeboot.py processes
The way killl_existing() is implemented seems very ugly. We identify a
"pxeboot.py" process by reading /proc and send it SIGKILL. What this in
particular does, is to keep the dhcpd service running. The code does
something, but it's unclear whether this is at all the best way.

The best way to run the tool is following the usage example in the README.md,
which does a `podman run --rm --replace --name marvell-tools...`.
If you do that, podman takes care that no concurrent dhcpd processes are
running.

If you don't follow that, then the tool cannot magically know which
processes on the system would interfere and should be stopped. The
caller must handle that, or use the suggested way via `podman run`.

This code seems unnecessary. Remove it.