pxeboot: support configuring static IP address and gateway on secondary interface
The secondary interface is expected to be connected to an interface on
the provisioning host, which runs DHCP.

However, it's not clear that this DHCP server keeps running. To have
basic connectivity even if the DHCP server is stopped, allow the
provisioning host to configure a static IP address on the secondary
interface "enP2p2s0-dpu-secondary".

Note that we still do DHCP in parallel. If we get a DHCP lease, we will
use that one in addition to the static IP address.

There is a problem with this. Now, both profiles "enP2p2s0-dpu-secondary"
and "enP2p3s0-dpu-host" can define static IP addresses and a gateway.
Also, "enP2p2s0-dpu-secondary" gets the better route metric, so that
gateway will be used to reach the internet. Is that gonna work? Is that
right? That depends on the environment. On the other hand, if the host
(on enP2p3s0-dpu-host and 172.131.100.1) runs CoreOS, it's also unclear
that we can reach the internet that way. So, whether this makes sense,
depends. The user may not opt-in to this.