pxeboot: detect yum repo url during pxeboot
During kickstart we might not have network to fetch the beaker URL.
However, we probably have from the node that runs pxeboot.

Try to detect the URL there. The relevant part is to detect the right
RHEL version, for that read "/mnt/marvel_dpu_iso/media.repo".

If we fail to read the URL, the kickstart script will also try to find
an URL.