pxeboot: rework wait_for_boot() to ping for static IP address
Previously, we were pinging for one of the IP addresses from the DHCP
range. Note that during PXE boot, the DPU would also acquire a DHCP
lease, but at that time the installation is still going on and not
what we wait for.

Hence, there was a sleep(1000) first, to ensure that we finished
installing the machine and booted the installed system. However, this
timeout is (intentionally) long, so that we don't accidentally complete
too early.

Instead, the kickstart also generates a NetworkManager profile with a
static IP address 172.131.100.100 (which is also not in the DHCP range).
So ping for that instead, and when we see it we know that our installed
system is up.

We can now ping earlier and will complete as soon as the DPU is ready.