all: use host.local.run() instead of run()
host.Host's run() implementation does conceptually the same thing as
common_dpu.run(). It however aims to do it the best way. Whether it
succeeds at that, can be questioned, but the point is that it's an
attempt of such an implementation that shall be reused and better (or
not worse) than what you could re-implement.

Use host.Host.run(), and don't reimplement it.

The main benefit is that it aims to do useful logging without extra
effort. It's arguably better than what common_dpu.run() did previously
as it does not wrap lines and shows the full output.

Also use shlex.quote() at various places where we paste string arguments
into a shell script.