pxeboot: add option to wait after starting HTTP/TFTP/DHCP services
For debugging and installing the DPU manually, it can be useful to use
pxeboot command to start the HTTP/TFTP/DHCP service and then just wait.

Add an option to prompt at that point. The user can then manually
connect to the DPU via the serial port and select the what to boot.