pxeboot: install "Minicom" command on DPU
One major purpose of the marvell-tools container is to use
it for manual debugging and access the terminal console.

For one, it has minicom installed (which CoreOS might not have). Also,
it places a ".minirc.dfl" file to disable flow control. That makes it
convenient to use.

Still, every time we have to type `minicom -D /dev/ttyUSB0`, and that
command is also not in the bash history. For convenience, add a
`Minicom` script that does this. It's really just to do less typing, now
`M<TAB>` suffices.