pxeboot: add shortcut flag "-H" for "--host-setup-only"
The "--host-setup-only" is useful to run every time we reboot
the host. It can be necessary to restore internet connectivity for the
DPU by enabling NAT and configuring a static IP address.

However, the flag is hard to remember or type. Instead, add a "-H"
shortcut which is easier to use.