Containerfile: add octep_cp_agent in arm64 container
On the DPU we need the octep_cp_agent user space program. Build it in
the arm64 container so we can conveniently run it on the DPU. Also add a
"/usr/bin/run_octep_cp_agent" command to the container which does all
the necessary steps necessary for setting up the DPU side.

On the DPU, during pxeboot (via kickstart) create a script
"/usr/bin/run_octep_cp_agent" to run the command from the container.
Also add a "octep_cp_agent.service" systemd service. For now this
service is automatically enabled on boot. Eventually, we will probably
not want this and let the dpu-operator run the agent somehow. For now,
having it run automatically via systemd is useful.