pxeboot: disable "NetworkManager-wait-online.service" on the DPU
"NetworkManager-wait-online.service" tends to hang and fail after timeout.
That causes a delay during boot and a bogus error message.

That is because we install two NetworkManager connection profiles that
autoconnect and do DHCP. That is for the interfaces connected to the
provisioning host and to the DPU's host. Often the DHCP servers on those
machines are not running (e.g. the DPU's host is usually a CoreOS and the
DHCP server was only there during the pxeboot step, but not after a
reboot). As a result, those profiles linger in "activating" state and
delay "NetworkManager-wait-online.service" until timeout.

Maybe the NetworkManager should be configured better (maybe using
"ipv4.required-timeout" setting?). But as it is, it's mostly fine until
proven otherwise.

Just disable "NetworkManager-wait-online.service" to avoid the problems
we have with this.

"NetworkManager-wait-online.service" has a certain purpose and makes
sense to enable by default. It's fine to disable it, in cases where it's
not the best solution.