Containerfile: install "tcpdump" and "git-core" in marvell-tools container
"tcpdump" is a useful debugging tool. Especially, since it is not installed
on CoreOS nodes by default. While we can always install it when running the
container with `dnf install tcpdump`, it seems useful to ship it out of
the box.

Also, we don't need all of "git", "git-core" is sufficient.