Containerfile: install "tcpdump", "unzip" and "git-core" in marvell-tools container
"tcpdump" is a useful debugging tool. Especially, since it is not installed
on CoreOS nodes by default. While we can always install it when running the
container with `dnf install tcpdump`, it seems useful to ship it out of
the box.

Also, we don't need all of "git", "git-core" is sufficient.

Also, CoreOS doesn't have unzip installed either. Sadly, tar does not
index the files while zip does. So it can be useful to share large
directores in zip. And it can be useful to have an unzip tool at hand.

Of course, the container is a functional CentOS Stream, so you can also
dnf-install packages. It's however convenient to pre-ship some useful
packages.