pxeboot: always create SSH keys also on CoreOS host
The pxeboot command has a "--host-mode" argument. For CoreOS, we do less
modifications, because a CoreOS is immutable and certain things don't
make sense.

However, the "/root" home directory is also persistet on CoreOS. It
makes sense to generate a SSH key there, and install it on the DPU.
The benefit is that you can SSH from the (CoreOS) host into the DPU,
which is commonly at IP address 172.131.100.100.