pxeboot: wait and log serial console output
After we select the right boot menu entry, the pxeboot command only
waits for the newly installed DPU to become ready.

In the background, the DPU will start PXE booting, and downloading
installation data via TFTP and HTTP. In the terminal, we see logging
messages from the HTTP server, but that may not be sufficient for
debugging.

Previously, we would no longer read the serial console at that point. So
if there is a problem and installation doesn't start, it's unclear why.

Avoid that, by keep reading the serial console for another minute and
log the output.

While it's a bit cumbersome, this way we can reconstruct from the
logfile every byte that was received.