common_dpu: improve logging by using ktoolbox.host.Host.run_in_thread()
Use FutureThread and Host.run() (Host.run_in_thread()) instead of
multiprocessing.Process. This results in nicer logging, and is supposed
to do the right thing in any case. If it doesn't, we would want to find
that bug in ktoolbox and fix our code, because we already use that at
other places too. And if it does the right thing, then use it.

Also see the logging output previously:

  2025-02-06 18:35:21.899 DEBUG   [th:139844862478144]: cmd[4;localhost]: call 'killall in.tftpd'
  2025-02-06 18:35:21.899 INFO    [th:139844862478144]: running /usr/sbin/dhcpd -f -cf /etc/dhcp/dhcpd.conf -user dhcpd -group dhcpd
  2025-02-06 18:35:21.906 DEBUG   [th:139844862478144]: cmd[4;localhost]: └──> 'killall in.tftpd': failed (exit 1); err='in.tftpd: no process found\n'

Note that the "running /usr/sbin/dhcpd" line is actually printed by
another process, but that is unclear as the thread-id in the logging was
cached and we see the outdated value after fork().

Also, we saw:

  2025-02-06 18:48:24.490 INFO    [th:140039096514368]: running /usr/sbin/in.tftpd -s -B 1468 -L /var/lib/tftpboot
  2025-02-06 18:48:24.493 INFO    [th:140039096514368]: Result:

  71

which does not a good job at printing the results of the command. This
could be improved by carefully crafing a logging line. But Host.run()
already does a better job here:

  2025-02-06 18:49:58.855 DEBUG   [th:139811879827008]: cmd[6;localhost]: call '/usr/sbin/in.tftpd -s -B 1468 -L /var/lib/tftpboot'
  ...
  2025-02-06 18:49:58.858 DEBUG   [th:139811879827008]: cmd[6;localhost]: └──> '/usr/sbin/in.tftpd -s -B 1468 -L /var/lib/tftpboot': failed (exit 71)

(also, note that the thread-id only in the latter example is correct and
shows a thread different from the main thread).

Yes, using an additional thread to control the process is more
expensive, but that is not a concern here.