pxeboot: don't install kernel via default-extra-packages
Support for Marvell DPU was merged to rhel-9 (9.6) kernel [1]
and is available in kernel-5.14.0-563.el9 and newer.

As such, we don't need any extra packages anymore. Drop this.

Note that support for this was not backported to rhel-9.5 and older
kernels. It was decided, that we only support rhel-9.6 and newer on the
DPU.

[1] https://gitlab.com/redhat/centos-stream/src/kernel/centos-stream-9/-/merge_requests/6271