pxeboot: use `podman run --pull newer` for running octep_cp_agent
pxeboot installs as octep_cp_agent.service systemd service, which
runs /usr/bin/run_octep_cp_agent script. That script is generated
by kickstart, and it runs a podman container via

  podman run --pull always ...

This fails, if the DPU happens to have no internet access, which can
easily happen if the gateway that we NAT through is not setup.

Instead, use "--pull newer" which tries to pull the image, but proceeds
taking the old one, if it fails to do so.