pxeboot: also with host-mode "coreos" write hosts entry and use NetworkManager
The host-mode parameter is supposed to potentially configure the
host differently, based on whether it is RHEL or CoreOS.

The two differences so far were:

  - add "dpu" entry to "/etc/hosts" only on RHEL (not CoreOS).

  - on RHEL, create a NetworkManager connection profile
    "eno4-marvell-dpu", while on CoreOS only add the address with
    iproute2.

Especially the second behavior is problematic. Even if we would not want
to mess with CoreOS, on CoreOS there will be autogenerated "Wired
connection N", which just does not the right thing for the eno4
interface. Also, even on CoreOS it would be good to have a persisted
connection profile that is still used after reboot.

Also, "/etc" is persisted on CoreOS. While a bit unorthodox to modify
them, it seems better than not to.

So, do the same for CoreOS and RHEL.

I think theoretically there could still be a use for the host-mode
parameter. Altough, for the moment it has no more effect.