pxeboot: write hosts entry and use NetworkManager also with host-mode "coreos"
The host-mode parameter is supposed to potentially configure the
host differently, based on whether it is RHEL or CoreOS.

The two differences so far were:

  - add "dpu" entry to "/etc/hosts" only on RHEL (not CoreOS).

  - on RHEL, create a NetworkManager connection profile
    "eno4-marvell-dpu", while on CoreOS only add the address with
    iproute2.

Especially the second point is not great for CoreOS. Even if we want to
minimize our changes to CoreOS, on CoreOS there will be autogenerated
NetworkManager profile "Wired connection N". This enabled DHCP and makes
no sense to use automatically on eno4 device. Also, even on CoreOS it
would be good to have a persisted connection profile that is still used
after reboot.

Also, "/etc" is persisted on CoreOS. While a bit unorthodox to modify
them, it seems better than not to.

So, do the same for CoreOS and RHEL.

Add a "ephermal" mode that avoids any persistent changes. But that one
is never chosen by default.