common_dpu: exit if services unexpectedly terminated
Logging an error is good, however, if the program continues it will
flood the output with messages from the serial console. So it is hard to
find this error message in the log.

Instead, abort the program at that point. There really is no expectation
that the process can continue without dhcpd/tftpd running.