common_dpu: workaround common failure to run dhcpd server
On CoreOS we sometimes find an address

    inet 192.168.122.101/32 scope global vip
       valid_lft forever preferred_lft forever

on br-ex interface. Note the address label "vip".

This also cause legacy `ifconfig` tool to fail, as in the past address
labels were used for detecting secondary IP addresses.

This also causes dhcpd to fail with "Error getting hardware address for "vip"".

Work around that by always deleting and readding the address (without
the label).

This seems to work well enough.