common_dpu: log error if command during run_process() fails
We use common_dpu.run_process() to run the dhcpd/tftpd servers.
They are not supposed to exit, unless we cancel (terminate) them
at the end.

Set check_success callback, to convert an exit without cancellation
to an error. Also, log an error message if the process terminates
in such failure.