pxeboot: add option "--octep-cp-agent-service-disable" to not enable systemd service
The tool will always create the "octep_cp_agent.service" systemd
service which can run octep_cp_agent.

By default, the service is enabled.

However, there is work in progress where the Marvell VSP of dpu-operator
will run the octep_cp_agent program. In that case, we will no longer
want to enable the service unconditionally. For now, add an option to
not enable the service.

Later, when the dpu-operator commonly runs the service, we will flip the
default and don't enable the service by default.