common_dpu: adjust packaged_file() to use common pattern
I think the preferred way to resolve relative paths from the
installation directory is this one.

First, detect the basedir based on __file__. This needs to be
done at the time when loading the module, since it might be
a relative file name, and os.getcwd() can change. We record
the detected cwd,basedir for later. Finally, use it to resolve
relative names.