pxeboot: add separate "ip-link-up.service" to setup network interfaces
To workaround RHEL-90248, we need to ensure that the SDP interfaces are
set up early on.

Previously, that was done by the "exec_octep_cp_agent" script, which is
inside the "marvell-tools" container, which is run by the
"run_octep_cp_agent" script, which is run by the
"octep_cp_agent.service". In short, this was previously done by the
"octep_cp_agent.service".

We soon will want to no longer enable "octep_cp_agent.service" by
default, but we still need the workaround. Hence, move it to the
separate "ip-link-up.service" service and "ip-link-up.sh" script.

Also, this allows to enable/disable the workaround separately, via
`systemctl enable/disable ip-link-up.service`. We thus no longer need
EXEC_OCTEP_CP_AGENT_SET_UP, which was anyway inconvenient to configure.