pxeboot: workaround broken ISO from earlier download
The pxeboot command downloads the ISO from a HTTP location. If the file
already exists, it is not re-downloaded.

However, if the file is corrupted, we may not immediately notice. In
that case, we only fail later, after trying to mount the iso and copying
out files. Something like:

   2025-09-05 14:23:21.067 DEBUG   [th:140289498121792]: cmd[12;localhost]: cancelled: terminate process 16335
   2025-09-05 14:23:21.067 DEBUG   [th:140289498121792]: cmd[12;localhost]: └──> '/usr/sbin/in.tftpd -s -B 1468 -L /var/lib/tftpboot': [CANCELLED] success [forced] [cancelled]
   2025-09-05 14:23:21.069 ERROR   [th:140289514180096]: FATAL ERROR:
   Traceback (most recent call last):
     File "/usr/local/lib/python3.9/site-packages/ktoolbox/common.py", line 2707, in run_main
       exit_code = main_fcn()
     File "/marvell-octeon-10-tools/./pxeboot.py", line 480, in main
       setup_tftp()
     File "/marvell-octeon-10-tools/./pxeboot.py", line 394, in setup_tftp
       shutil.copy(
     File "/usr/lib64/python3.9/shutil.py", line 427, in copy
       copyfile(src, dst, follow_symlinks=follow_symlinks)
     File "/usr/lib64/python3.9/shutil.py", line 287, in copyfile
       copyfileobj(fsrc, fdst)
     File "/usr/lib64/python3.9/shutil.py", line 205, in copyfileobj
       buf = fsrc_read(length)
   OSError: [Errno 5] Input/output error

What is worse, subsequent runs will continue to fail until we manually
remove the broken file from disk.

Instead. If the file is an HTTP URL and it already exists, then perform
an early (naive) check that we can read the ISO. If the file seems
unusable, try to re-download the file.