pxeboot: in kickstart support ignoring TLS errors for dnf-install
We will want to install some packages from a HTTPS URL with self signed
certificates. `dnf install` is hard to convince to accept that. So it is
more cumbersome.

Note that we also handle http:// URLs, because those get automatically
redirected to https:// (but then the certificate is untrusted).

Note that we only do this special handling if the caller opts in by
setting the URL to "untrusted:http://..." or "untrusted:https://...". We
don't want to ignore certificate errors by default.