pxeboot: move copy_kickstart() function to IsoKindRhel class
Next we will unify the code. This function thus should be part of
IsoKindRhel class. Move it in a first commit, with no other changes.

The point is to do the trivial move of the code first, so that the diff
later gets with the real changes is easier to understand.