pxeboot: refactor copy_kickstart() to IsoKind.setup_http_files()
This rename will make sense next, when we add another IsoKind
implementation that does a different setup.

Also, we don't want to pass that many parameters to a function. The
entre point of RunContext is to provide a (almost) immutable object as a
container for many parameters, which then can be passed around freely
(as functions only will read the value, not change state).