pxeboot: write serial output to "/tmp/pxeboot-serial.YYYYMMDD-hhmmss.ffffff.log"
During pxeboot it is rather useful to be able to look at the output on
the serial console. If you combine that with "--console-wait" option, the
tool stays longer connected to the serial console and will keep logging.

All the content was already logged. And you could use
"scripts/pxeboot-parse-log.sh" script to parse it from the log. However,
this is still cumbersome.

As an alternative debugging help, write the output to a logfile.