pxeboot: while waiting for boot continue reading serial console
During installation, we only need the serial console initially, when
selecting the boot menu.

After that, we usually want to close the serial console, because
otherwise the user cannot connect another `minicom` process to the
console.

In some cases, it can however be useful to let the pxeboot be attached
to the serial console. That is what the "--console-wait" option is for.
With it, we keep the serial console connected, which has the benefit
that we read and log whatever we see on the console.

However, previously, we would always hang for the "--console-wait" time
to only read the console. In particular, we wouldn't notice when the
machine completed booting fast.

Now, intertwine reading the serial console and waiting for boot to
complete. The benefit is that if we set a large "--console-wait"
timeout, then we still will quit once the machine is booted.