pxeboot: wait until we can SSH into the host
PXE boot differs from regular boot from ISO, because we anyway must
unpack an ISO and share data on HTTP/TFTP in a suitable manner.
At that point, we always do some configuration regarding what we boot.

In particular, we also register our own authenticated SSH key, so that
we can access the host afterwards.

Previously, when booting we waited until the static IP address was
pingable. That is not very reliable and it does not yet mean that the
host is ready. We also cannot wait to ping the DHCP addresses, because
those might be available even during installation.

Instead, ping all the IP addresses that possibly are configured by the
booted host. If we get a reply, we try to SSH into it. Only then we
consider the boot completed.