pxeboot: stay connected to Serial console until end
We only need the serial console initially to select the boot menu.
However, for debugging it is rather useful to stay connected. In that
case, the tool will continue to read from the console, log the messages
and write them to "/host/tmp/pxeboot-serial*log".

The downside is that while pxeboot is connected, we cannot otherwise
connect with minicom. On the other hand, if we close the serial port,
we will loose messages.

Previously, the tool stayed connected for "--console-wait" seconds,
which defaulted to 60 seconds. No, since we write the output to
file, it seems more useful by default to stay connected.

Interpret "--console-wait 0" to mean to stay connected indefinitely.
Also, this is the new default.

Also, honor SIGUSR1 to close the serial port. So by default we will stay
connected and log everything to "/tmp". If you want to manually connect
a minicom session while the pxeboot command is still running, you can
send a SIGUSR1 to close the console and take over.