pxeboot: no longer install test kernel for reproducing RHEL-90248
There was a lot of back and forth with RHEL-90248.

- first, there was a severe flakiness (RHEL-90248) which prevented us
  from enabling CI.
- then, a workaround was found by running a "ip-link-up.service"
  service in marvell-octeon-10-tools.
- then, an upstream kernel patch was identified which might fix
  RHEL-90248.
- at that point, I was unable to reproduce the original issue. Disabling
  the workaround "ip-link-up.service" seems to have mostly no effect.
  This is despite the kernel versions are similar (but not the same),
  the firmware and cp-agent versions were the same and the BIOS settings
  should be the same.
- as a result, RHEL-90248 was closed as CANT-REPRODUCE.
- still, marvell-octeon-10-tools was updated to install a test kernel
  via extra packages. That test kernel contains the upstream patch above.

Now, revert that test kernel again. Let's see how that goes.

If we don't notice any bad failures, the issues is really not
reproducible. If we notice bad failures, then the patch is a fix for
RHEL-90248. Given that there were already difficulties to reproduce
RHEL-90248, I expect to be the former and the official kernel will work
just as fine. Let's see.