pxeboot: encode URL in downloaded ISO file
It is very convenient that pxeboot supports HTTP/HTTPS URLs for
the ISO. The tool will download them to "{chroot_path}/root"
and re-use the file, if it already exists from an earlier run.

Imagine you have the URL to a RHCOS discovery ISO, which was created for
an AssistedInstaller infraenv. The file name of that URL is always
"full.iso".

Previously, the stored filename would just be the name in the URL. So
for different URLs a "rhel-iso-full.iso" file could match. That is a
problem.

Instead, add a hash of the URL to the filename. That way, different URLs
result in different files, and we don't accidentally take a stale file
from disk.