pxeboot: adjust regex for different MAC address for selecting boot entry
I followed Kiet's email with "howto_fix_mac_addresses.txt" to configure
fixed MAC addresses. I ended up with:

  5) BOARD-MAC-ADDRESS-ID4 (0x000fb70655d4)

This means the UEFI boot menu looks now different.

Adjust the pattern to match either the old or the new MAC address. Note
that the new pattern gives some room, so we could configure a slightly
different MAC address on the various DPUs that we have.

This is not a great solution, but it unblocks it from now. In the
future, we probably should improve the detection of the right boot
entry.