pxeboot: add option to boot from primary/secondary DPU device
Add an option "--dpu-dev" to specify the PXE boot interface on the DPU
side. This can be "primary", "secondary" or the MAC address.

This determines which interface the pxeboot command will select to PXE
boot.

So far, we were always using the primary interface. With single cluster
setup for dpu-operator, it may make more sense to use the secondary
network. In that case, the primary network (enP2p3s0) is connected to
the OCP network (192.168.122.0/24), while the secondary network
(enP2p2s0) is used for PXE boot and later for the external network.
Such a setup requires a re-wiring the network connections compared to
how we usually setup the cluster.

Also, the DHCP server now only replies to requests from the expected
hardware address. This is relevant, because if the tools runs on the
secondary network, that network probably already has another DHCP server
running.  The DHCP server from the pxeboot command will only serve
PXEClient requests from the DPU and not disrupt the network otherwise.

It now might also make sense to run the pxeboot command on another host.
In that case, run with "-P" so that the tool stops after starting DHCP,
TFTP and HTTP servers. Then, manually access the serial console and
select the PXE boot entry in the UEFT bios. At least for manual use,
this allows to install the DPU using the secondary network, even if the
DPU host is not connected to that network.