pxeboot: detect host-mode based on ISO kind
Detecting the host most is useful. When we install a RHCOS, we assume
that the caller does not want to permanently modify the system that runs
pxeboot command. If we have a RHCOS ISO, assume ephemeral configuration
of the host.

Note that we also detect the ISO kind, so we must do the ISO detection
as very first.