pxeboot: add "--dhcp-restricted" option to handle random MAC addresses
Add a new --dhcp-restricted option (auto/yes/no) to control whether the
DHCP server restricts requests to a specific MAC address.

With "yes" (restricted mode), the DHCP server only responds to the
specific DPU MAC address that was detected, which is useful when running
on a network with an existing DHCP server.

With "no" (unrestricted mode), the DHCP server responds to any client,
serving a range of addresses (172.131.100.10-20) in addition to the
fixed address for the known host.

With "auto" (the default), the behavior is determined automatically
based on the well-known MAC address pattern that shows up in the Marvell
DPU's UEFI boot menu. If the MAC matches 80:aa:99:88:77:6[67], it's
considered random/unstable and unrestricted mode is used. Otherwise,
restricted mode is used.

This is important because the DPU may be configure to use use randomized
MAC addresses that change on each boot, making restricted mode
ineffective.