pxeboot: reset DPU after detecting MAC address with "--prompt"
If the user did "--prompt", then maybe they just want to start the DHCP
service, etc.

To detect the MAC address, we already reset the DPU once. Usually, we
later continue with PXE booting, and will reset again. With "--prompt"
that might not happen. In that case, explicitly reset the DPU, so that
it might boot into the exepcted system.