pxeboot: exit gracefully with CTRL+C during prompt
Once point of the prompt is to run the DHCP/TFTP/HTTP server.
The user can choose to continue booting (that is, telling the
DPU to PXE boot) or abort with CTRL+C.

In the latter case, convert the KeyboardInterrupt to a regular
exit, to not get a backtrace in the console output.