pxeboot: rework parsing of MAC addresses from boot menu
On the DPU, we can configure up to 4 ports. The UEFI boot menu
consequently looks different. Rework the parsing for that.

Note that the "primary" interface (the RJ45 port) is always the last. So
first are the N configured "secondary" interfaces (starting at index
zero) and then at index N is the primary interface. With "--dpu-dev" we
can specify either the MAC address, the (zero based) index, or the
special words "primary" and "secondary".  "secondary" is the first
secondary and an alias for "0". If you need other (secondary) interfaces
use "1", "2, or "3".