pxeboot: track Serial instance in RunContext
The benefit is that the RunContext can access the Serial port too. That
means, we can implement a getter method on the RunContext that looks up
the status in the console.

Another benefit is that the RunContext knows whether currently a serial
connection is open.

- calling another create will fail if a serial connection already exists
- accessing via RunContext.serial_get() will fail, if no connection was
  created.

This is nice, because if you get an exception, you know you do something
wrong.